Created the page to see profile of users. #129
